+++ differt/src/differt/geometry/_paths.py
@@ -133,6 +133,16 @@ class Paths(eqx.Module):
             )
             warnings.warn(msg, UserWarning, stacklevel=2)
 
+    @property
+    def shape(self) -> tuple[int, ...]:
+        """
+        Return the batch shape of the paths.
+
+        Returns:
+            The shape of paths' batch dimensions.
+        """
+        return self.vertices.shape[:-2]
+
     def reshape(self, *batch: int) -> Self:
         """
         Return a copy with reshaped paths' batch dimensions to match a given shape.
@@ -291,7 +301,7 @@ class Paths(eqx.Module):
     @property
     def order(self) -> int:
         """The length (i.e., number of vertices) of each individual path, excluding start and end vertices."""
-        return self.objects.shape[-1] - 2
+        return self.path_length - 2
 
     @property
     def num_valid_paths(self) -> int | Int[Array, " "]:
@@ -339,6 +349,40 @@ class Paths(eqx.Module):
             mask = self.confidence.reshape(-1) >= self.confidence_threshold
             return objects[mask, ...]
         return objects
+
+    def masked(self) -> Self:
+        """Return a flattened copy of this object that only keeps valid paths.
+
+        TODO: document.
+        """
+        batch_len = len(self.vertices.shape) - 2
+        self = jax.tree.map(lambda arr: arr.reshape(-1, *arr.shape[batch_len:]), self)
+
+        if self.mask is not None:
+            mask = self.mask.reshape(-1)
+        elif self.confidence is not None:
+            mask = self.confidence.reshape(-1) >= self.confidence_threshold
+        else:
+            mask = slice(None, None, None)
+
+        return eqx.tree_at(
+            lambda p: (
+                p.vertices[mask, ...],
+                p.objects[mask, ...],
+                None,
+                p.interaction_types[mask, ...],
+                None,
+            ),
+            self,
+            (
+                self.vertices,
+                self.objects,
+                self.mask,
+                self.interaction_types,
+                self.confidence,
+            ),
+            is_leaf=lambda x: x is None,
+        )
 
     @eqx.filter_jit
     def multipath_cells(
@@ -390,7 +434,8 @@ class Paths(eqx.Module):
         if self.mask is not None:
             mask = jnp.moveaxis(self.mask, axis, -1)
         else:
-            mask = jnp.moveaxis(self.confidence >= self.confidence_threshold, axis, -1)
+            # Looks like Pyright is not able to infer this
+            mask = jnp.moveaxis(self.confidence >= self.confidence_threshold, axis, -1)  # type: ignore[reportOperatorIssue]
         *partial_batch, last_axis = mask.shape
 
         return _cell_ids(mask.reshape(-1, last_axis)).reshape(partial_batch)
@@ -568,11 +613,8 @@ class SBRPaths(Paths):
         )
 
     def plot(self, **kwargs: Any) -> PlotOutput:
-        backend = kwargs.pop(
-            "backend", None
-        )  # TODO: check if kwargs may not cause issues
-        with reuse(backend=backend) as output:
+        with reuse(**kwargs, pass_all_kwargs=True) as output:
             for order in range(self.order + 1):
-                self.get_paths(order).plot(**kwargs)
+                self.get_paths(order).plot()
 
         return output